@@ -2,75 +2,69 @@
   "name": "peer-review",
   "version": "0.0.0",
   "license": "MIT",
-  "scripts": {
-    "ng": "ng",
-    "start": "webpack-dev-server --port=4200",
-    "build": "webpack",
-    "meow": "npm run build & npm run start",
-    "test": "karma start ./karma.conf.js",
-    "lint": "ng lint",
-    "e2e": "protractor ./protractor.conf.js",
-    "pree2e": "webdriver-manager update --standalone false --gecko false --quiet"
-  },
+  "scripts": {},
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/animations": "^6.0.2",
+    "@angular/common": "^6.0.2",
+    "@angular/compiler": "^6.0.2",
+    "@angular/core": "^6.0.2",
+    "@angular/forms": "^6.0.2",
+    "@angular/http": "^6.0.2",
+    "@angular/platform-browser": "^6.0.2",
+    "@angular/platform-browser-dynamic": "^6.0.2",
+    "@angular/router": "^6.0.2",
     "core-js": "^2.4.1",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.1.0",
     "zone.js": "^0.8.19"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.5",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.6.3",
+    "@angular/cli": "6.0.3",
+    "@angular/compiler-cli": "^6.0.2",
+    "@angular/language-service": "^6.0.2",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "autoprefixer": "^7.2.3",
+    "@types/node": "^10.1.2",
+    "autoprefixer": "^8.5.0",
     "bootstrap": "^4.1.1",
-    "circular-dependency-plugin": "^4.2.1",
+    "circular-dependency-plugin": "^5.0.2",
     "codelyzer": "^4.0.1",
     "copy-webpack-plugin": "^4.1.1",
     "css-loader": "^0.28.1",
     "cssnano": "^3.10.0",
-    "exports-loader": "^0.6.3",
-    "file-loader": "^1.1.11",
-    "html-webpack-plugin": "^2.29.0",
+    "exports-loader": "^0.7.0",
+    "file-loader": "^1.1.5",
+    "html-webpack-plugin": "^3.2.0",
     "istanbul-instrumenter-loader": "^3.0.0",
-    "jasmine-core": "~2.8.0",
+    "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
     "jquery": "^3.3.1",
     "karma": "~2.0.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^2.0.0",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine-html-reporter": "^1.1.0",
     "less-loader": "^4.0.5",
     "node-sass": "^4.9.0",
+    "popper.js": "^1.14.3",
     "postcss-import": "^11.0.0",
     "postcss-loader": "^2.0.10",
     "postcss-url": "^7.1.2",
-    "protractor": "~5.1.2",
+    "protractor": "~5.3.2",
     "raw-loader": "^0.5.1",
-    "sass-loader": "^6.0.6",
+    "sass-loader": "^7.0.1",
     "source-map-loader": "^0.2.0",
-    "style-loader": "^0.13.1",
+    "style-loader": "^0.21.0",
     "stylus-loader": "^3.0.1",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
+    "ts-node": "~6.0.3",
+    "tslint": "~5.10.0",
+    "typescript": "^2.7.2",
     "uglifyjs-webpack-plugin": "^1.1.5",
-    "url-loader": "^0.6.2",
-    "webpack": "~3.10.0",
-    "webpack-dev-server": "~2.11.0"
+    "url-loader": "^1.0.1",
+    "webpack": "~4.8.3",
+    "webpack-cli": "^2.1.3",
+    "webpack-dev-server": "~3.1.4"
   }
 }
