@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {HttpClient, HttpHeaders, HttpParams} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs/index';
 import {CourseModel} from '../models/course.model';
 
 @Injectable()
@@ -16,15 +16,9 @@ export class Api {
 
     })
   };
+
   // 'Access-Control-Allow-Origin': '*'}
-  // private httpOptions = {
-  //   headers: new HttpHeaders({
-  //     'Access-Control-Allow-Origin': '*'
-  //     // ,
-  //     // 'Authorization': 'authkey',
-  //     // 'userid': '1'
-  //   })
-  // };
+
 
   constructor(public http: HttpClient) {
 
@@ -34,16 +28,26 @@ export class Api {
     return this.baseUrl + url + `?TokenData=${this.TokenData}&UserId=${this.UserId}`;
   }
 
+  private getPureUrl(url: string = ''): string {
+    return this.baseUrl + url;
+  }
+
+
   public get(url: string = ''): Observable<any> {
     return this.http.get(this.getUrl(url)); // .subscribe(data => console.log(data));
   }
 
   public post(url: string = '', data: any = {}): Observable<any> {
     console.log('json', data);
-    return this.http.post(this.getUrl(url), data); // .subscribe(data => console.log(data));
+    return this.http.post(this.getUrl(url), data, this.httpOptions); // .subscribe(data => console.log(data));
   }
 
   public put(url: string = '', data: any = {}): Observable<any> {
     return this.http.put(this.getUrl(url), data, this.httpOptions); // .subs
+  }
+
+  public postWithoutParams(url: string = '', data: any = {}): Observable<any> {
+    console.log('json', data);
+    return this.http.post(this.getPureUrl(url), data, this.httpOptions); // .subscribe(data => console.log(data));
   }
 }
