@@ -1,12 +1,12 @@
 import {Component, OnInit} from '@angular/core';
-import {Subscription} from 'rxjs/Subscription';
+import {Subscription} from 'rxjs/index';
 import {ReviewService} from '../../../shared/services/review.service';
 import {SolutionDataService} from '../solution-data.service';
-import {SolutionModel} from '../../../../common/models/solution.model';
 import {NgForm} from '@angular/forms';
 import {ReviewModel} from '../../../../common/models/review.model';
-import {TaskDataService} from '../../task/task-data.service';
 import {CriteriaModel} from '../../../../common/models/criteria.model';
+import {ReviewCriteriaModel} from '../../../../common/models/review-criteria.model';
+import {TaskDataService} from '../../task/task-data.service';
 
 @Component({
   selector: 'peer-review-review-add',
@@ -18,17 +18,35 @@ export class ReviewAddComponent implements OnInit {
   solutionId: string;
   criteria: CriteriaModel[];
 
-  constructor(private reviewService: ReviewService, private taskDataService: TaskDataService) {
+  constructor(private reviewService: ReviewService, private solutionDataService: SolutionDataService,
+              private taskDataService: TaskDataService
+  ) {
 
   }
 
   ngOnInit() {
     this.criteria = this.taskDataService.getTask().CriteriaCollection;
+    this.solutionId = this.solutionDataService.getSolution().Id;
   }
 
   onSubmit(form: NgForm) {
+    console.log(form.value);
+    let reviewCriteriaCollection = [];
+    this.criteria.forEach((c, i) => {
+        const points = form.value['point-' + i];
+        const reviewCriteria = new ReviewCriteriaModel(c.Title, points ? points : c.MaxPoint, c.Id);
+        reviewCriteriaCollection.push(reviewCriteria);
+      }
+    );
+    const review = new ReviewModel(
+      window.localStorage.userId,
+      this.solutionId,
+      new Date(),
+      reviewCriteriaCollection
+    );
+    console.log(review);
+    console.log(JSON.stringify(review));
 
-    // const solution = new ReviewModel(title, description, this.taskId, new Date(), false);
-    // this.sub = this.solutionService.createSolution(solution).subscribe(data => console.log(data));
+    this.sub = this.reviewService.createReview(review).subscribe(data => console.log(data));
   }
 }
