@@ -1,8 +1,7 @@
-import {Component, OnInit} from '@angular/core';
-import {SolutionModel} from '../../../../common/models/solution.model';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {SolutionDataService} from '../solution-data.service';
 import {ReviewModel} from '../../../../common/models/review.model';
-import {Subscription} from 'rxjs/Subscription';
+import {Subscription} from 'rxjs/index';
 import {ReviewService} from '../../../shared/services/review.service';
 
 @Component({
@@ -10,7 +9,7 @@ import {ReviewService} from '../../../shared/services/review.service';
   templateUrl: './solution-reviews.component.html',
   styleUrls: ['./solution-reviews.component.scss']
 })
-export class SolutionReviewsComponent implements OnInit {
+export class SolutionReviewsComponent implements OnInit, OnDestroy {
 
   reviews: ReviewModel[];
   sub: Subscription;
@@ -24,7 +23,12 @@ export class SolutionReviewsComponent implements OnInit {
     this.sub = this.reviewService.getReviewListBySolution(solutionId).subscribe(data => {
       this.reviews = data;
       this.isLoaded = true;
+      console.log(data);
     });
+  }
+
+  ngOnDestroy(): void {
+    if (this.sub) this.sub.unsubscribe();
   }
 
 }
