@@ -6,7 +6,8 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./navbar.component.scss']
 })
 export class NavbarComponent implements OnInit {
-
+  token = window.localStorage.token;
+  userId = window.localStorage.userId;
   constructor() { }
 
   ngOnInit() {
