@@ -1,9 +1,10 @@
 import {Injectable} from '@angular/core';
 import {Api} from '../../../common/core/api';
 import {SolutionModel} from '../../../common/models/solution.model';
-import {Observable} from 'rxjs/Observable';
+
 import {HttpClient, HttpParams} from '@angular/common/http';
 import {ReviewModel} from '../../../common/models/review.model';
+import {Observable} from 'rxjs/index';
 
 @Injectable()
 export class ReviewService extends Api {
@@ -13,13 +14,10 @@ export class ReviewService extends Api {
   }
 
   createReview(review: ReviewModel): Observable<ReviewModel> {
-    this.params = new HttpParams({
-      fromObject: JSON.parse(JSON.stringify(review))
-    });
-    return this.post(`Review`);
+    return this.post(`Review`, review);
   }
 
-  getReviewById(token: string, reviewId: number): Observable<ReviewModel> {
+  getReviewById( reviewId: string): Observable<ReviewModel> {
     return this.get(`reviews/${reviewId}`);
   }
 
